Hand selection added. Logging view now animates according to hand.
Hand is selected in MainActivity.

Aware and logging only start after Begin button is pressed in
LoggingActivity. They only stop after the Done button is pressed.

Tap indicators in TapView display sequentially depending on the
selectedHand string passed to LoggingActivity. This is done via an
AsyncTask which updates the position of each indicator every 2 seconds.

After the 5 points are displayed, the Done button is displayed.